Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,6 +5,6 @@ int main() {
     scanf("%d",&n1);
     scanf("%d",&n2);
     scanf("%d",&n3);
-    printf("Average: %.2f",(n1+n2+n3)/3);
+    printf("Average: %.2f",(float)(n1+n2+n3)/3);
     return 0;
 }
